+++ packages/node/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support endpoint configs and specifying headers for network endpoints (#2511)
+
 ### Fixed
 - Fixed timestamp can be undefined in some network blocks, it should return undefined (#2513)